Fix V1 blockers for range routing, docs upload auth, and nav sync
