+++ src/app/range/drill-library/page.tsx
@@ -1,0 +1,5 @@
+import { RangeWorkspace } from "@/components/range/RangeWorkspace";
+
+export default function DrillLibraryPage() {
+  return <RangeWorkspace view="drill-library" />;
+}
