+++ src/app/range/drill-performance/page.tsx
@@ -1,0 +1,5 @@
+import { RangeWorkspace } from "@/components/range/RangeWorkspace";
+
+export default function DrillPerformancePage() {
+  return <RangeWorkspace view="drill-performance" />;
+}
