+++ src/app/range/hit-factor/page.tsx
@@ -1,0 +1,5 @@
+import { RangeWorkspace } from "@/components/range/RangeWorkspace";
+
+export default function HitFactorPage() {
+  return <RangeWorkspace view="hit-factor" />;
+}
