+++ src/app/range/log-drill/page.tsx
@@ -1,0 +1,5 @@
+import { RangeWorkspace } from "@/components/range/RangeWorkspace";
+
+export default function LogDrillPage() {
+  return <RangeWorkspace view="log-drill" />;
+}
