+++ src/app/range/log-session/page.tsx
@@ -1,0 +1,5 @@
+import { RangeWorkspace } from "@/components/range/RangeWorkspace";
+
+export default function LogSessionPage() {
+  return <RangeWorkspace view="log-session" />;
+}
