+++ src/app/range/session-history/page.tsx
@@ -1,0 +1,5 @@
+import { RangeWorkspace } from "@/components/range/RangeWorkspace";
+
+export default function SessionHistoryPage() {
+  return <RangeWorkspace view="session-history" />;
+}
